Feat/#10

See merge request ormee1/server!17

+++ src/main/java/com/ormee/server/dto/LectureResponseDto.java
@@ -21,6 +21,6 @@ public class LectureResponseDto {
     private LocalTime endTime;
     private LocalDateTime dueTime;
     private List<QuizListDto> quizList;
-    private Integer activeQuizCount;
+    private Long activeQuizCount;
     private boolean messageAvailable;
 }
